Moving help component modules for better reusability
